Better implementation that allows overriding classes of HystrixServoMetricsPublisherCommand to be less-aware of implementation details
